[evals] Move trace chat dataset into trace module
