+++ lib/levanter/src/levanter/data/text/trace_chat.py
@@ -2,16 +2,25 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ast
+import functools
 import json
 import re
 from collections.abc import Mapping, Sequence
 from dataclasses import dataclass
-from typing import Any, Literal, TypedDict
+from typing import Any, Literal, TypedDict, cast
 
+import equinox as eqx
+import jax
+import jax.numpy as jnp
 import numpy as np
+from haliax import Axis
 
 from levanter.data._preprocessor import BatchProcessor
-from levanter.data.text.examples import LossLabelSpec
+from levanter.data.dataset import MappedAsyncDataset
+from levanter.data.packing import GreedyPrepackedDataset
+from levanter.data.sharded_datasource import ShardedDataSource
+from levanter.data.text.examples import GrugAttentionMask, LabeledLmExample, LossLabelSpec
+from levanter.store.cache import CacheOptions, TreeCache
 from levanter.tokenizers import MarinTokenizer
 
 
@@ -28,6 +37,10 @@ _FIRST_CUSTOM_TRACE_LABEL = 32
 class ProcessedTraceChatDict(TypedDict):
     input_ids: np.ndarray
     loss_labels: np.ndarray
+
+
+def _single_cpu_sharding() -> jax.sharding.SingleDeviceSharding:
+    return jax.sharding.SingleDeviceSharding(jax.local_devices(backend="cpu")[0])
 
 
 _TEXT_TOOL_CALL_PATTERN = re.compile(r"<tool_call>\s*(.*?)\s*</tool_call>", re.DOTALL)
@@ -222,6 +235,84 @@ class TraceChatEvaluationFormat:
 
     def loss_label_spec(self) -> LossLabelSpec:
         return loss_label_spec_for_trace_tags(self.loss_tags)
+
+
+class TraceChatDataset(MappedAsyncDataset[tuple[ProcessedTraceChatDict, ProcessedTraceChatDict], LabeledLmExample]):
+    """A dataset that yields packed trace examples with exclusive loss labels."""
+
+    def __init__(
+        self,
+        cache: TreeCache[ProcessedTraceChatDict],
+        Pos: Axis,
+        max_segments_per_example: int = 64,
+        slice_strategy: Literal["left", "right", "raise"] = "left",
+        block_cross_document_attention: bool = True,
+    ):
+        self.packed: GreedyPrepackedDataset[ProcessedTraceChatDict] = GreedyPrepackedDataset(
+            cache.jagged_array_tree(),
+            Pos.size,
+            max_segments_per_example=max_segments_per_example,
+            slice_strategy=slice_strategy,
+        )
+        self.Pos = Pos
+        self.block_cross_document_attention = block_cross_document_attention
+
+        sharding = _single_cpu_sharding()
+
+        @functools.partial(eqx.filter_jit)
+        def _create_trace_example(e: tuple[ProcessedTraceChatDict, ProcessedTraceChatDict]) -> LabeledLmExample:
+            example, seg_ids = e
+            tokens = example["input_ids"]
+            segment_ids = seg_ids["input_ids"]
+
+            labels = jnp.roll(example["loss_labels"], -1, axis=-1)
+            next_segment_ids = jnp.roll(segment_ids, -1, axis=-1)
+            same_segment_next = (segment_ids == next_segment_ids) & (segment_ids >= 0)
+            same_segment_next = same_segment_next.at[-1].set(False)
+            labels = jnp.where(same_segment_next, labels, 0)
+
+            attn_mask = GrugAttentionMask.causal()
+            if block_cross_document_attention:
+                attn_mask = attn_mask.with_segment_ids(segment_ids)
+
+            out = LabeledLmExample(tokens=tokens, loss_labels=labels, attn_mask=attn_mask)
+            out = jax.lax.with_sharding_constraint(out, sharding)
+            return out
+
+        super().__init__(self.packed, _create_trace_example)
+
+
+def build_trace_chat_dataset_cache(
+    cache_dir: str,
+    source: ShardedDataSource[dict],
+    trace_format: TraceChatEvaluationFormat,
+    tokenizer: MarinTokenizer,
+    options: CacheOptions = CacheOptions.default(),
+) -> TreeCache[ProcessedTraceChatDict]:
+    processor = trace_format.build_preprocessor(tokenizer)
+    return cast(
+        TreeCache[ProcessedTraceChatDict], TreeCache.build_or_load(cache_dir, source, processor, options=options)
+    )
+
+
+def dataset_for_trace_chat_format(
+    trace_format: TraceChatEvaluationFormat,
+    Pos: Axis,
+    cache: TreeCache[ProcessedTraceChatDict],
+    *,
+    block_cross_document_attention: bool = True,
+) -> TraceChatDataset:
+    pack = trace_format.pack
+    if pack == "pad":
+        raise NotImplementedError("Padding mode not yet implemented.")
+    max_segments = 64 if pack is True or pack is None else int(pack)
+    return TraceChatDataset(
+        cache,
+        Pos,
+        max_segments_per_example=max_segments,
+        slice_strategy=trace_format.slice_strategy,
+        block_cross_document_attention=block_cross_document_attention,
+    )
 
 
 class TraceChatProcessor(BatchProcessor[dict, dict]):
